Final refactor before merge.
