+++ .rubocop.yml
@@ -29,6 +29,9 @@ Metrics/PerceivedComplexity:
 RSpec/ExampleLength:
   Enabled: false
 
+RSpec/MultipleExpectations:
+  Enabled: false
+
 RSpec/NestedGroups:
   Enabled: false
